Add files via upload
Fixed bug in calculator. Cleaned up UX and made hitting enter take you to the calculation instead of hitting 0 to do that. That choice was a bit strange by me... Well fixed now! Will be adding the ability to square single numbers, as well as raise numbers to a power in future. Will also still be adding history showing the numbers that led to answers, and GUI.